+++ src/material-examples/tab-nav-bar-basic/tab-nav-bar-basic-example.css
@@ -1,3 +1,4 @@
 .example-action-button {
-  margin-bottom: 8px;
+  margin-top: 8px;
+  margin-right: 8px;
 }
