+++ src/material/tabs/tab-nav-bar/tab-nav-bar.scss
@@ -1,22 +1,18 @@
 @import '../tabs-common';
 @import '../../core/style/variables';
 
-// Wraps the bar containing the anchors
-.mat-tab-nav-bar {
-  @include tab-header;
-}
+@include paginated-tab-header;
 
 .mat-tab-links {
-  position: relative;
-  display: flex;
+  @include paginated-tab-header-item-wrapper;
+}
 
-  [mat-align-tabs='center'] & {
-    justify-content: center;
-  }
+.mat-ink-bar {
+  @include ink-bar;
+}
 
-  [mat-align-tabs='end'] & {
-    justify-content: flex-end;
-  }
+.mat-tab-link-container {
+  @include paginated-tab-header-container;
 }
 
 // Wraps each link in the header
@@ -42,14 +38,8 @@
   }
 }
 
-
 @media ($mat-xsmall) {
   .mat-tab-link {
     min-width: 72px;
   }
-}
-
-// Styling for the ink bar that displays near the activated anchor
-.mat-ink-bar {
-  @include ink-bar;
 }
