+++ src/material/tabs/tab-nav-bar/tab-nav-bar.spec.ts
@@ -23,6 +23,7 @@ describe('MatTabNavBar', () => {
         TabLinkWithNgIf,
         TabLinkWithTabIndexBinding,
         TabLinkWithNativeTabindexAttr,
+        TabBarWithInactiveTabsOnInit,
       ],
       providers: [
         {provide: MAT_RIPPLE_GLOBAL_OPTIONS, useFactory: () => globalRippleOptions},
@@ -199,7 +200,17 @@ describe('MatTabNavBar', () => {
 
       expect(inkBar.hide).toHaveBeenCalled();
     });
+
   });
+
+  it('should hide the ink bar if no tabs are active on init', fakeAsync(() => {
+    const fixture = TestBed.createComponent(TabBarWithInactiveTabsOnInit);
+    fixture.detectChanges();
+    tick(20); // Angular turns rAF calls into 16.6ms timeouts in tests.
+    fixture.detectChanges();
+
+    expect(fixture.nativeElement.querySelector('.mat-ink-bar').style.visibility).toBe('hidden');
+  }));
 
   it('should clean up the ripple event handlers on destroy', () => {
     let fixture: ComponentFixture<TabLinkWithNgIf> = TestBed.createComponent(TabLinkWithNgIf);
@@ -365,3 +376,15 @@ class TabLinkWithTabIndexBinding {
   `
 })
 class TabLinkWithNativeTabindexAttr {}
+
+
+@Component({
+  template: `
+    <nav mat-tab-nav-bar>
+      <a mat-tab-link *ngFor="let tab of tabs" [active]="false">Tab link {{label}}</a>
+    </nav>
+  `
+})
+class TabBarWithInactiveTabsOnInit {
+  tabs = [0, 1, 2];
+}
